requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
fastapi==0.117.1
ffmpy==0.6.1
filelock==3.19.1
fsspec==2025.9.0
gradio==5.46.1
gradio_client==1.13.1
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
portalocker==3.2.0
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pynput==1.8.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
ruff==0.13.1
safehttpx==0.1.6
-e git+https://github.com/facebookresearch/segment-anything-2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4#egg=SAM_2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.36.0
websockets==15.0.1
